+++ Frends.Files.Copy/Frends.Files.Copy/Definitions/Options.cs
@@ -56,5 +56,16 @@ public class Options
     /// </summary>
     /// <example>FileExistsAction.Overwrite</example>
     public FileExistsAction IfTargetFileExists { get; set; }
-}
 
+    /// <summary>
+    /// Whether to throw an error on a failure to copy a file.
+    /// This option is independent of the <see cref="IfTargetFileExists"/> option with the value of <see cref="FileExistsAction.Throw"/> and will affect handling of all errors during the copy process.
+    /// If set to false, task will continue executing despite files failing to copy, otherwise execution will stop at the first failure.
+    /// 
+    /// Note: When set to false and <see cref="IfTargetFileExists"/> is set to <see cref="FileExistsAction.Throw"/>, the task will continue copying other files despite a failure.
+    /// Behaviour of <see cref="FileExistsAction.Overwrite"/> and <see cref="FileExistsAction.Rename"/> will not be affected by this setting.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFail { get; set; }
+}