+++ Frends.Files.Delete/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [1.1.0] - 2025-03-19
+## [1.2.0] - 2025-03-19
 ### Changed
 - Update packages:
   Microsoft.Extensions.FileSystemGlobbing  7.0.0 -> 9.0.3
@@ -12,6 +12,10 @@
   MSTest.TestFramework                     2.2.8  -> 3.8.3
   nunit                                    3.12.0 -> 4.3.2
   NUnit3TestAdapter                        3.17.0 -> 5.0.0
+
+## [1.1.0] - 2025-03-13
+### Fixed
+- Fixed bug where regex patterns were not handled correctly in file matching.
 
 ## [1.0.0] - 2023-02-13
 ### Added